sop_generator/requirements.txt
streamlit>=1.28.0
autogen-agentchat>=0.2.0
openai>=1.3.0
python-docx>=0.8.11
PyPDF2>=3.0.1
pandas>=2.1.0
openpyxl>=3.1.0
Pillow>=10.0.0
reportlab>=4.0.4
streamlit-ace>=0.1.1
pdf2image>=1.17.0
pytesseract>=0.3.10 